updated styles for tablet and laptop
